Introduce ScribePurge class
