+++ includes/cache/SquidUpdate.php
@@ -116,7 +116,7 @@ class SquidUpdate {
 		// wikia change start
 		global $wgPurgeSquidViaScribe;
 		if( $wgPurgeSquidViaScribe == true ) {
-			SquidUpdate::ScribePurge( $urlArr );
+			ScribePurge::purge( $urlArr );
 			return;
 		}
 
@@ -231,78 +231,5 @@ class SquidUpdate {
 	 */
 	static function expand( $url ) {
 		return wfExpandUrl( $url, PROTO_INTERNAL );
-	}
-
-	/**
-	 * wikia function
-	 * @
-	 * @access public
-	 * @static
-	 */
-	static function ScribePurge( $urlArr ) {
-		global $wgEnableScribeReport, $wgCityId;
-		wfProfileIn( __METHOD__ );
-		$key = 'varnish_purges';
-
-		if ( empty($wgEnableScribeReport) ) {
-			wfProfileOut( __METHOD__ );
-			return true;
-		}
-
-		try {
-			foreach ( $urlArr as $url ) {
-				if ( !is_string( $url ) ) {
-					throw new MWException( 'Bad purge URL' );
-				}
-				$url = SquidUpdate::expand( $url );
-				$method = self::getPurgeCaller();
-
-				wfDebug( "Purging URL $url from $method via Scribe\n" );
-				wfDebug( "Purging backtrace: " . wfGetAllCallers( false ) . "\n" );
-				$data = json_encode(
-					array(
-						'url' => $url,
-						'time' => time(),
-						'method' => $method,
-					)
-				);
-				WScribeClient::singleton($key)->send($data);
-
-				// log purges using SFlow (BAC-1258)
-				Wikia\SFlow::operation('varnish.purge', [
-					'city' => $wgCityId,
-					'url' => $url,
-					'method' => $method,
-				]);
-			}
-		}
-		catch( TException $e ) {
-			Wikia::log( __METHOD__, 'scribeClient exception', $e->getMessage() );
-		}
-
-		wfProfileOut( __METHOD__ );
-	}
-
-	/**
-	 * Return the name of the method (outside of internal code) that triggered purge request
-	 *
-	 * @return bool|string method name
-	 */
-	private static function getPurgeCaller() {
-		// analyze the backtrace to log the source of purge requests
-		$backtrace = wfDebugBacktrace();
-		$method = '';
-
-		while($entry = array_shift($backtrace)) {
-			// ignore "internal" classes
-			if (empty($entry['class']) || in_array($entry['class'], [__CLASS__, 'WikiPage', 'Article', 'Title'])) {
-				continue;
-			}
-
-			$method = $entry['class'] . ':' . $entry['function'];
-			break;
-		}
-
-		return $method;
 	}
 }
